ㅅㅎ

+++ src/main/java/com/springboot/biz/free/answer/FreeAnswerController.java
@@ -4,15 +4,19 @@ package com.springboot.biz.free.answer;
 import com.springboot.biz.free.board.FreeQuestion;
 import com.springboot.biz.free.board.FreeQuestionRepository;
 import com.springboot.biz.free.board.FreeQuestionService;
+import com.springboot.biz.user.HUser;
+import com.springboot.biz.user.HUserSerevice;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
+
+import java.security.Principal;
 
 @Controller
 @RequiredArgsConstructor
@@ -21,6 +25,7 @@ public class FreeAnswerController {
 
     private final FreeAnswerService freeAnswerService;
     private final FreeQuestionService freeQuestionService;
+    private final HUserSerevice hUserSerevice;
 
 
 
@@ -29,64 +34,70 @@ public class FreeAnswerController {
     @PostMapping("/create/{frboSeq}")
     public String createFreeAnswer(@PathVariable("frboSeq") Integer frboSeq,
                                    @Valid FreeAnswerForm freeAnswerForm,
-                                   BindingResult bindingResult, Model model) {
+                                   BindingResult bindingResult, Model model , Principal principal) {
         FreeQuestion freeQuestion = freeQuestionService.getFreeQuestion(frboSeq);
+        HUser hUser = hUserSerevice.getUser(principal.getName());
 
         if (bindingResult.hasErrors()) {
             model.addAttribute("freeQuestion", freeQuestion);
             return "free/freeQuestion_detail";
         }
 
-        freeAnswerService.freeAnswerCreate(freeQuestion, freeAnswerForm.getFrboAnsContent(), null);
+        freeAnswerService.freeAnswerCreate(freeQuestion, freeAnswerForm.getFrboAnsContent(), null,hUser);
 
         return String.format("redirect:/freequestion/detail/%s", frboSeq);
     }
 
 
 
+
+
+
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/modify/{frboAnSeq}")
     public String modifyFreeAnswer(@PathVariable("frboAnSeq") Integer frboAnSeq,
-                                   @Valid FreeAnswerForm freeAnswerForm, BindingResult bindingResult) {
+                                   @RequestParam("frboAnsContent") String frboAnsContent,
+                                   Principal principal) {
 
-        // 유효성 검사 실패 시 원래 페이지로 이동
-        if (bindingResult.hasErrors()) {
-            return "redirect:/freequestion/detail/" + freeAnswerService.getFreeAnswer(frboAnSeq).getFreeQuestion().getFrboSeq();
-        }
-
-        // 기존 댓글 가져오기
         FreeAnswer freeAnswer = freeAnswerService.getFreeAnswer(frboAnSeq);
 
-        // 댓글 수정
-        freeAnswerService.modify(freeAnswer, freeAnswerForm.getFrboAnsContent());
+        if (!freeAnswer.getFrboAnsAuthor().getUsername().equals(principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정 권한이 없습니다.");
+        }
 
-        // 수정 후 원래 질문 상세 페이지로 리디렉트
+        // 서비스로 수정 위임
+        freeAnswerService.modify(freeAnswer, frboAnsContent);
+
+        // 원글로 이동
         return "redirect:/freequestion/detail/" + freeAnswer.getFreeQuestion().getFrboSeq();
     }
 
-
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/delete/{frboAnSeq}")
-    public String deleteFreeAnswer(@PathVariable("frboAnSeq") Integer frboAnSeq) {
+    public String deleteFreeAnswer(@PathVariable("frboAnSeq") Integer frboAnSeq, Principal principal) {
         FreeAnswer freeAnswer = freeAnswerService.getFreeAnswer(frboAnSeq);
         Integer questionSeq = freeAnswer.getFreeQuestion().getFrboSeq(); // 질문 번호 가져오기
+
+        if(!freeAnswer.getFrboAnsAuthor().getUsername().equals(principal.getName())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST,"삭제 권한이 없습니다." );
+        }
 
         freeAnswerService.delete(freeAnswer); // 댓글 삭제
 
         return "redirect:/freequestion/detail/" + questionSeq; // 질문 상세 페이지로 이동
     }
 
+
     @PostMapping("/reply/{frboAnSeq}")
     public String createReply(@PathVariable("frboAnSeq") Integer frboAnSeq,
                               @Valid FreeAnswerForm freeAnswerForm,
-                              BindingResult bindingResult, Model model) {
+                              BindingResult bindingResult,Principal principal) {
         FreeAnswer parentAnswer = freeAnswerService.getFreeAnswer(frboAnSeq);
         FreeQuestion freeQuestion = parentAnswer.getFreeQuestion();
+        HUser hUser = hUserSerevice.getUser(principal.getName());
 
-        if (bindingResult.hasErrors()) {
-            model.addAttribute("freeQuestion", freeQuestion);
-            return "free/freeQuestion_detail";
-        }
 
-        freeAnswerService.freeAnswerCreate(freeQuestion, freeAnswerForm.getFrboAnsContent(), parentAnswer);
+        freeAnswerService.freeAnswerCreate(freeQuestion, freeAnswerForm.getFrboAnsContent(), parentAnswer,hUser);
 
         return String.format("redirect:/freequestion/detail/%s", freeQuestion.getFrboSeq());
     }
